requirements.txt
openai>=1.95.1
PyYAML>=6.0.3